20170619_x64_完整修訂版

+++ CommandTest/LaserAdjust.cpp
@@ -204,14 +204,14 @@ void CLaserAdjust::ListRefresh()
 	size_t nCount = ((CCommandTestDlg*)pMain)->a.LaserAdjust.size();
 	for (size_t i = 0; i < nCount; i++)
 	{
-		m_ListLaserAdjust.InsertItem(i, NULL);
+		m_ListLaserAdjust.InsertItem((int)i, NULL);
 		(i > 8) ? StrBuff.Format(_T("0%d"), i + 1) : StrBuff.Format(_T("00%d"), i + 1);
-		m_ListLaserAdjust.SetItemText(i, 0, StrBuff);
+		m_ListLaserAdjust.SetItemText((int)i, 0, StrBuff);
 		StrBuff.Format(L"%d", ((CCommandTestDlg*)pMain)->a.LaserAdjust.at(i).LaserMeasureHeight);
-		m_ListLaserAdjust.SetItemText(i, 1, StrBuff);
+		m_ListLaserAdjust.SetItemText((int)i, 1, StrBuff);
 	}
     size_t ListnCount = m_ListLaserAdjust.GetItemCount();
-	m_ListLaserAdjust.EnsureVisible(ListnCount - 1, FALSE);//使List中一項可見(如滾動條向下滾)
+	m_ListLaserAdjust.EnsureVisible((int)(ListnCount - 1), FALSE);//使List中一項可見(如滾動條向下滾)
 }
 //顯示視窗時設定
 void CLaserAdjust::OnShowWindow(BOOL bShow, UINT nStatus)
